Fixed id generation on Rust side. Physics dynamic actors are using that.
Still crashing with car due to not being done yet.

+++ Thunder/physics-handler/src/lib.rs
@@ -290,7 +290,7 @@ impl PhysicsHandler {
         is_dynamic: bool,
         transform: PxTransform,
         geometry: &impl Geometry,
-        id: u64,
+        id: usize,
         static_shape_filter_data: u32,
     ) {
         if is_dynamic {
@@ -310,7 +310,7 @@ impl PhysicsHandler {
                 physx_sys::PxCollection_add_mut(
                     self.raw_ptr_holder.physx_collection,
                     actor.as_mut_ptr(),
-                    id as usize,
+                    id,
                 );
                 physx_sys::PxSerialization_complete_mut(
                     self.raw_ptr_holder.physx_collection,
@@ -353,7 +353,7 @@ impl PhysicsHandler {
                 physx_sys::PxCollection_add_mut(
                     self.raw_ptr_holder.physx_collection,
                     actor.as_mut_ptr(),
-                    id as usize,
+                    id,
                 );
                 physx_sys::PxSerialization_complete_mut(
                     self.raw_ptr_holder.physx_collection,
